// src/main/webapp/app/core/layout/sidenav/sidenav-item/sidenav-item.interface.ts
export class SidenavItem {
    name: string;
    icon?: string;
    routeOrFunction?: any;
    parent?: SidenavItem;
    subItems?: SidenavItem[];
    position?: number;
    pathMatchExact?: boolean;
    badge?: string;
    badgeColor?: string;
    type?: 'item' | 'subheading';
    customClass?: string;
    //campo para la validacion de usuario OJO
    authenticationGear?: string;
}